feat: Create new dotnet project for media transfer API
It is planned to refactor the APIs used to copy uploads from internet to intranet environment to improve the process by preventing lost uploads and minimising processing delay.

To prepare for the refactoring, create a new project to host the new APIs since the new APIs are for internal consumption only. The old transfer APIs will be removed from the MediaLibrary.Internet.Web project in future commits.

- Create MediaLibrary.Internet.Api.csproj for new API
- Add API controller with preliminary API methods that return mock responses
   - Get list of items for transfer
   - Get file contents
   - Delete an item (called after item has been transferred to intranet env)
- Use Swashbuckle to generate OpenAPI specification and provide Swagger UI