fix: intranet: remove "event" string replacement during processing
The string replacement was probably added to work around "event" being a reserved word in C#. However, this could erroneously modify other content in the JSON as the implementation does not check where the string "event" appears.

In addition, the workaround is unnecessary as reserved words can be used in variable names by prefixing it with @ symbol. JSON deserialisation will work as intended as "@keyword" compiles to "keyword" in the CIL.