fix: Try harder to reduce image file size for image content analysis
Cognitive Services APIs have a 4MB size limit for input images. Improve the algorithm used to fit image file size within the maximum limit. This happens in three stages:
  1. If original image resolution exceeds 25MP, scale the image down to around 24MP
  2. Repeatedly scale image dimensions by 95%, up to 10 iterations
  3. Repeatedly lower JPEG quality from 80 to 50 in steps of 5

If at any point during the process the current image is less than 4MB in size, stop the process and return the current image.

Also fixes an infinite loop that could happen with the previous algorithm when it fails to generate a smaller image of acceptable file size.