Fix bugs in declaration section
If user completed the declaration for a draft but decides to cancel it, a new draft created thereafter will not require the declaration to be done. This is because the declaration status is not reset when the draft is cancelled. To fix, reset the status when entering the step and make sure the input is controlled by FormContext.

Also fix various styling and alignment issues with the section.

Finally, remove up extra whitespace in the error message when declaration is not completed.